+++ src/taskpane/pages/dashboard.tsx
@@ -4,7 +4,7 @@ import data from "../../utils/data";
 import ReferenceList, { bib } from "../components/ReferenceList";
 import SearchField from "../components/SearchField";
 import CiteSupport from "../../utils/citesupport";
-import { citationByIndexInterface } from "../../utils/cite-interface";
+import { citationByIndexInterface } from "../../utils/citeproc-types";
 import WordApi from "../../utils/word-api";
 
 interface dashboardProps {
